+++ Todo-service/src/main/java/com/school/repository/StudentRepository.java
@@ -3,5 +3,5 @@ package com.school.repository;
 import com.school.entities.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<Student, Integer> {
+public interface StudentRepository extends JpaRepository<Student, Long> {
 }
